LayoutTests/imported/w3c/web-platform-tests/contenteditable/plaintext-only-expected.txt


PASS plaintext-only is an accepted attribute value for contenteditable
PASS plaintext-only can be assigned to contenteditable dynamically
